Use floats for movement calculation.
Haven't set the values yet, but got everything set up so that Vel uses
floats instead of ints. Will still be using ints for collision, but
floats for integrating position.

+++ src/BlasterShot.cpp
@@ -5,7 +5,7 @@
 
 using namespace std;
 
-BlasterShot::BlasterShot(int startX, int startY, int velocityX, int velocityY) {
+BlasterShot::BlasterShot(int startX, int startY, float velocityX, float velocityY) {
     hitbox.x = startX;
     hitbox.y = startY;
     hitbox.w = TILE_WIDTH;
